requirements.txt
# Core dependencies
azure-core==1.35.0
azure-ai-inference==1.0.0b9

python-multipart==0.0.20
fastapi==0.116.1
uvicorn==0.35.0

opencv-python==4.12.0.88
numpy==2.0.2

minio==7.2.16